# backend/src-generated/api/openapi.yaml
openapi: 3.0.3
info:
  description: |
    API REST para la plataforma de búsqueda inversa de empleo
    de la Municipalidad de Providencia, Chile.

    ## Roles del sistema
    - **admin** — Departamento de Empleo Municipal
    - **talento** — Vecinos/as de Providencia
    - **empresa** — Empresas empleadoras

    ## Autenticación
    Todos los endpoints protegidos requieren un token JWT en el header:
    `Authorization: Bearer <token>`
  title: ProviEmplea API
  version: 1.0.0
servers:
- description: Servidor de desarrollo
  url: http://localhost:3000/api/v1
...
paths:
  /auth/login:
    post:
      description: Autentica un usuario y devuelve un JWT válido por 24h
      operationId: authLoginPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_auth_login_post_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
          description: Login exitoso
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Credenciles inválidas
      summary: Login de usuario
      tags:
      - Auth
      x-eov-operation-handler: controllers/AuthController
  /auth/register/talento:
    post:
      description: La cuenta queda en esta Pendiente hasta que un admin la valide
      operationId: authRegisterTalentoPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_auth_register_talento_post_request"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
          description: "Registro exitoso, cuenta pendiente de validación"
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Datos inválidos
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: El correo ya está registrado
      summary: Registro de vecino/talento
      tags:
      - Auth
      x-eov-operation-handler: controllers/AuthController
  /auth/register/empresa:
    post:
      operationId: authRegisterEmpresaPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_auth_register_empresa_post_request"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
          description: Empresa registrada exitosamente
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Rut o correo ya registrado
      summary: Registro de empresa
      tags:
      - Auth
      x-eov-operation-handler: controllers/AuthController
...
components:
  schemas:
    SuccessResponse:
      example:
        data: "{}"
        success: true
        message: Operación exitosa
      properties:
        success:
          example: true
          type: boolean
        message:
          example: Operación exitosa
          type: string
        data:
          type: object
      type: object
    ErrorResponse:
      example:
        success: false
        message: Descripción del error
      properties:
        success:
          example: false
          type: boolean
        message:
          example: Descripción del error
          type: string
      type: object
...
    _auth_login_post_request:
      properties:
        correo:
          example: maria@gmail.com
          format: email
          type: string
        password:
          example: MiPassword123!
          type: string
      required:
      - correo
      - password
      type: object
    _auth_register_talento_post_request:
      properties:
        correo:
          example: vecino@gmail.com
          format: email
          type: string
        password:
          example: Password123!
          type: string
      required:
      - correo
      - password
      type: object
    _auth_register_empresa_post_request:
      properties:
        correo:
          example: rrhh@empresa.cl
          format: email
          type: string
        password:
          example: Password123!
          type: string
        rut_empresa:
          example: 76.123.456-7
          type: string
        nombre_empresa:
          example: Tech Solutions Spa
          type: string
      required:
      - correo
      - nombre_empresa
      - password
      - rut_empresa
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: Token JWT obtenido desde POST /auth/login
      scheme: bearer
      type: http